:recycle: Move getTransactionId into transactions module
