+++ test/transactions/utils/index.js
@@ -17,6 +17,7 @@ import {
 	getTimeWithOffset,
 	getTransactionBytes,
 	getTransactionHash,
+	getTransactionId,
 	prepareTransaction,
 } from '../../../src/transactions/utils';
 
@@ -34,8 +35,12 @@ describe('transaction utils', () => {
 			getTransactionBytes.should.be.type('function');
 		});
 
-		it('should have getTransactionBytes', () => {
+		it('should have getTransactionHash', () => {
 			getTransactionHash.should.be.type('function');
+		});
+
+		it('should have getTransactionId', () => {
+			getTransactionId.should.be.type('function');
 		});
 
 		it('should have prepareTransaction', () => {
